feat: Use CircleCI instead of Travis for tests

+++ tests/test_integration_all_rust.py
@@ -22,7 +22,7 @@ import psutil
 import requests
 import twisted.internet.base
 from autopush.db import (
-    DynamoDBResource, create_message_table, create_router_table
+    DynamoDBResource, create_message_table, get_router_table
 )
 from autopush.tests.test_integration import Client, _get_vapid
 from autopush.utils import base64url_encode
@@ -31,6 +31,7 @@ from Queue import Empty, Queue
 from twisted.internet import reactor
 from twisted.internet.defer import inlineCallbacks, returnValue
 from twisted.trial import unittest
+from typing import Optional
 
 app = bottle.Bottle()
 logging.basicConfig(level=logging.DEBUG)
@@ -41,7 +42,7 @@ root_dir = os.path.dirname(here_dir)
 
 DDB_JAR = os.path.join(root_dir, "ddb", "DynamoDBLocal.jar")
 DDB_LIB_DIR = os.path.join(root_dir, "ddb", "DynamoDBLocal_lib")
-DDB_PROCESS = None  # type: subprocess.Popen
+DDB_PROCESS = None  # type: Optional[subprocess.Popen]
 
 twisted.internet.base.DelayedCall.debug = True
 
@@ -252,18 +253,21 @@ def capture_output_to_queue(output_stream):
 def setup_dynamodb():
     global DDB_PROCESS
 
-    cmd = " ".join([
-        "java", "-Djava.library.path=%s" % DDB_LIB_DIR,
-        "-jar", DDB_JAR, "-sharedDb", "-inMemory"
-    ])
-    DDB_PROCESS = subprocess.Popen(cmd, shell=True, env=os.environ)
     if os.getenv("AWS_LOCAL_DYNAMODB") is None:
+        print "Starting new DynamoDB instance"
+        cmd = " ".join([
+            "java", "-Djava.library.path=%s" % DDB_LIB_DIR,
+            "-jar", DDB_JAR, "-sharedDb", "-inMemory"
+        ])
+        DDB_PROCESS = subprocess.Popen(cmd, shell=True, env=os.environ)
         os.environ["AWS_LOCAL_DYNAMODB"] = "http://127.0.0.1:8000"
+    else:
+        print "Using existing DynamoDB instance"
 
     # Setup the necessary tables
     boto_resource = DynamoDBResource()
     create_message_table(MESSAGE_TABLE, boto_resource=boto_resource)
-    create_router_table(ROUTER_TABLE, boto_resource=boto_resource)
+    get_router_table(ROUTER_TABLE, boto_resource=boto_resource)
 
 
 def setup_mock_server():
@@ -351,7 +355,9 @@ def setup_module():
 
 
 def teardown_module():
-    kill_process(DDB_PROCESS)
+    if DDB_PROCESS:
+        os.unsetenv("AWS_LOCAL_DYNAMODB")
+        kill_process(DDB_PROCESS)
     kill_process(CN_SERVER)
     kill_process(CN_MP_SERVER)
     kill_process(EP_SERVER)
